requirements.txt
################################################################################
# LEGDeeplab: Advanced Semantic Segmentation Framework
# Production-ready requirements file with detailed dependency management
################################################################################

# Core Deep Learning Framework
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Enhanced Performance (optional but recommended)
torch>=1.12.0+cu113; sys_platform == "linux"  # CUDA 11.3 optimized
torch>=1.12.0+cu116; sys_platform == "win32"  # Windows CUDA 11.6 optimized

# Essential Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
scikit-image>=0.19.0
scikit-learn>=1.1.0

# Image Processing & Computer Vision
Pillow>=8.3.2
opencv-python>=4.6.0.66
opencv-contrib-python>=4.6.0.66
albumentations>=1.2.0
imgaug>=0.4.0

# Data Handling & I/O
pandas>=1.4.0
h5py>=3.6.0
tqdm>=4.62.0
lmdb>=1.3.0

# Machine Learning Utilities
tensorboard>=2.8.0
mlflow>=1.23.0
wandb>=0.12.10
optuna>=2.10.0

# Visualization & Plotting
matplotlib>=3.5.0
seaborn>=0.11.2
plotly>=5.6.0
bokeh>=2.4.3

# Advanced Neural Network Operations
einops>=0.4.1
kornia>=0.6.0
timm>=0.6.0

# Model Optimization & Deployment
onnx>=1.11.0
onnxruntime>=1.11.0
onnxsim>=0.4.0
tensorrt>=8.4.0.0; sys_platform == "linux"  # GPU-accelerated inference

# Performance Analysis & Profiling
thop>=0.1.1-1  # For FLOPs calculation
fvcore>=0.1.5.post20220512  # Facebook Vision Core
memory-profiler>=0.61.0
line-profiler>=3.5.1

# Development & Code Quality
black>=22.1.0
flake8>=4.0.1
isort>=5.10.1
pre-commit>=2.17.0
pytest>=7.0.1
pytest-cov>=3.0.0

# Advanced Configuration Management
hydra-core>=1.1.2
omegaconf>=2.1.1
yacs>=0.1.8

# Distributed Training Support
torch-distributed>=0.12.0
deepspeed>=0.6.0
apex-cuda>=0.1  # NVIDIA Apex for mixed precision

# Advanced Mathematical Operations
numba>=0.55.0
cupy-cuda11x>=10.0.0; sys_platform == "linux"  # GPU-accelerated NumPy
cython>=0.29.28

# Specialized Computer Vision Libraries
detectron2 @ https://github.com/facebookresearch/detectron2/archive/main.zip
mmcv-full>=1.5.0  # OpenMMLab Computer Vision Foundation
mmdet>=2.25.0  # OpenMMLab Detection Toolbox
mmsegmentation>=0.25.0  # OpenMMLab Segmentation Toolbox

# Cloud & Storage Integration
boto3>=1.21.0  # AWS SDK
azure-storage-blob>=12.9.0  # Azure Blob Storage
google-cloud-storage>=2.1.0  # Google Cloud Storage

# Version Control & Experiment Tracking
gitpython>=3.1.27
dvc>=2.10.2
gdown>=4.4.0  # Google Drive downloader

# Security & Validation
pyyaml>=6.0
jsonschema>=4.4.0

# Advanced Logging & Monitoring
rich>=12.0.0
colorama>=0.4.4
loguru>=0.6.0

# Async & Parallel Processing
asyncio-mqtt>=0.11.0
concurrent-log-handler>=0.9.20

# Testing & Benchmarking
pytest-benchmark>=3.4.1
memory-profiler>=0.61.0

################################################################################
# Installation Instructions:
################################################################################

# Basic installation:
# pip install -r requirements.txt

# For development with extra tools:
# pip install -r requirements-dev.txt

# For production deployment:
# pip install -r requirements-prod.txt

################################################################################
# Optional Dependencies (install separately based on needs):
################################################################################

# For Jupyter Notebook Support:
# jupyter>=1.0.0
# jupyterlab>=3.3.0
# ipywidgets>=7.6.5

# For Web Deployment:
# flask>=2.0.3
# streamlit>=1.8.1
# gradio>=3.1.0

# For Mobile Deployment:
# tensorflow-lite>=2.8.0
# coremltools>=5.2.0; sys_platform == "darwin"

################################################################################
# Platform-specific recommendations:
################################################################################

# Linux:
# For optimal performance, use conda-forge packages:
# conda install -c conda-forge numpy scipy scikit-image pillow opencv

# Windows:
# Some packages may need to be installed from conda-forge:
# conda install -c conda-forge opencv pillow scikit-image

# macOS:
# For M1/M2 chips, ensure ARM64 compatibility:
# pip install --no-binary torch torchvision torchaudio torch

################################################################################
# Version Notes:
################################################################################

# Known compatibility issues:
# - albumentations < 1.1.0 has issues with newer OpenCV versions
# - scikit-image >= 0.19.0 requires newer SciPy
# - torch >= 1.12.0 has breaking changes in some APIs

# Recommended combinations:
# For RTX 30xx/40xx cards: torch>=1.12.0+cu113 or higher
# For older GPUs: torch>=1.10.0+cu111
# For CPU-only: torch>=1.12.0+cpu